+++ python/api_implementation.c
@@ -25,7 +25,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "python/python.h"
+#include "python/python_api.h"
 
 static struct PyModuleDef module_def = {
     PyModuleDef_HEAD_INIT,
